chore(shared) move schemas to their respective views
