+++ projects/@shared/views/favicon/favicon.view.vue
