+++ projects/@shared/views/google/google.view.vue
