+++ projects/@shared/views/linkedin/linkedin.view.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="slack">
+  <div class="linkedin">
     <panel-section title="Preview">
       <p v-if="!hasRequiredData">
         This page does not contain the necessary metadata to create a preview.
@@ -7,11 +7,11 @@
       <preview-iframe
         v-else
         :url="previewUrl"
-        iframeClass="slack__preview"
-        :loading-height="263"
+        iframeClass="linkedin__preview"
+        :loading-height="348"
       >
         <template v-slot:caption>
-          Preview based on <external-link href="https://slack.com/">slack.com</external-link>.
+          Preview based on <external-link href="https://linkedin.com/">linkedin.com</external-link>.
         </template>
       </preview-iframe>
     </panel-section>
@@ -30,21 +30,29 @@
         </properties-item>
       </properties-list>
     </panel-section>
+    <panel-section title="Resources">
+      <ul class="resource-list">
+        <li>
+          <external-link href="https://www.linkedin.com/post-inspector/">
+            LinkedIn Post Inspector
+          </external-link>
+        </li>
+        <li>
+          <external-link href="https://kinsta.com/blog/linkedin-debugger/">
+            Previews on LinkedIn (Post Inspector Tips)
+          </external-link>
+        </li>
+      </ul>
+    </panel-section>
   </div>
 </template>
 
 <script>
 import { computed, onMounted, ref, watch } from 'vue';
 import createAbsoluteUrl from '@shared/lib/create-absolute-url';
-import {
-  findAdditionalTwitterData,
-  findFavicons,
-  findImageDimensions,
-  findMetaContent,
-  findMetaProperty
-} from '@shared/lib/find-meta';
+import { findImageDimensions, findMetaContent, findMetaProperty } from '@shared/lib/find-meta';
 import getTheme from '@shared/lib/theme';
-import schema from '@shared/lib/schemas/slack-schema';
+import schema from './schema';
 
 import ExternalLink from '@shared/components/external-link';
 import PanelSection from '@shared/components/panel-section';
@@ -63,38 +71,27 @@ export default {
   setup: props => {
     const imageDimensions = ref({ height: undefined, width: undefined });
     const hasRequiredData = computed(() => (
-      (og.value.title || props.headData.head.title) &&
-      (og.value.description || headDescription.value)
+      Boolean(og.value.title) &&
+      Boolean(og.value.image) &&
+      Boolean(og.value.description) &&
+      Boolean(og.value.url)
     ));
     const og = computed(() => ({
       title: propertyValue('og:title'),
       type: propertyValue('og:type'),
-      description: propertyValue('og:description'),
-      site_name: propertyValue('og:site_name'),
       image: propertyValue('og:image'),
+      description: propertyValue('og:description'),
       url: propertyValue('og:url'),
-    }));
-    const headDescription = computed(() => (findMetaContent(props.headData.head, 'description')));
-    const favicon = computed(() => (
-      findFavicons(props.headData.head).length ? findFavicons(props.headData.head)[0].url : ''
-    ));
-    const additional = computed(() => ({
-      favicon: favicon.value,
-      twitterData: findAdditionalTwitterData(props.headData.head),
     }));
     const previewUrl = computed(() => {
       const params = new URLSearchParams();
-      params.set('additionalData', JSON.stringify(additional.value.twitterData));
-      params.set('description', og.value.description || headDescription.value);
-      params.set('favicon', additional.value.favicon);
+      params.set('title', og.value.title);
       params.set('image', og.value.image);
-      params.set('imageIsBig', imageDimensions.value.height > 201 && imageDimensions.value.width > 201);
-      params.set('siteName', og.value.site_name);
+      params.set('description', og.value.description);
+      params.set('url', og.value.url);
       params.set('theme', getTheme());
-      params.set('title', og.value.title || props.headData.head.title || 'Weblink');
-      params.set('url', props.headData.head.url);
-      params.set('validImage', imageDimensions.value.height > 0 && imageDimensions.value.width > 0);
-      return `/previews/slack/slack.html?${ params }`;
+      params.set('imageIsBig', imageDimensions.value.height >= 400 && imageDimensions.value.width >= 400);
+      return `/previews/linkedin/linkedin.html?${ params }`;
     });
     const metaData = computed(() => ([
       {
@@ -108,14 +105,6 @@ export default {
         required: true,
       },
       {
-        term: 'og:description',
-        value: og.value.description,
-      },
-      {
-        term: 'og:site_name',
-        value: og.value.site_name,
-      },
-      {
         term: 'og:image',
         value: og.value.image,
         image: {
@@ -126,19 +115,15 @@ export default {
         required: true,
       },
       {
+        term: 'og:description',
+        value: og.value.description,
+      },
+      {
         term: 'og:url',
-        value: absoluteUrl(og.value.url),
+        value: og.value.url,
         type: 'link',
         required: true,
       },
-      // Transform twitterData array and
-      // spread objects into metaData array.
-      ...additional.value.twitterData.map((item, index) => (
-        Object.entries(item).map(entry => ({
-          term: `twitter:${ entry[0] === 'label' ? entry[0] : 'data' }${ index + 1 }`,
-          value: entry[1],
-        }))
-      )).flat(),
     ]));
 
     const absoluteUrl = url => createAbsoluteUrl(props.headData.head, url);
@@ -158,9 +143,6 @@ export default {
     return {
       hasRequiredData,
       og,
-      headDescription,
-      favicon,
-      additional,
       previewUrl,
       metaData,
       schema,
@@ -177,7 +159,7 @@ export default {
 </script>
 
 <style>
-  .slack__preview {
+  .linkedin__preview {
     max-width: var(--preview-width);
   }
 </style>
