+++ projects/@shared/views/linkedin/schema.js
