+++ projects/@shared/views/open-search/schema.js
