+++ projects/@shared/views/slack/schema.js
