+++ projects/@shared/views/slack/slack.view.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="whatsapp">
+  <div class="slack">
     <panel-section title="Preview">
       <p v-if="!hasRequiredData">
         This page does not contain the necessary metadata to create a preview.
@@ -7,11 +7,11 @@
       <preview-iframe
         v-else
         :url="previewUrl"
-        iframeClass="whatsapp__preview"
-        :loading-height="122"
+        iframeClass="slack__preview"
+        :loading-height="263"
       >
         <template v-slot:caption>
-          Preview based on <external-link href="https://web.whatsapp.com/">web.whatsapp.com</external-link>.
+          Preview based on <external-link href="https://slack.com/">slack.com</external-link>.
         </template>
       </preview-iframe>
     </panel-section>
@@ -30,28 +30,21 @@
         </properties-item>
       </properties-list>
     </panel-section>
-    <panel-section title="Resources">
-      <ul class="resource-list">
-        <li>
-          <external-link href="https://stackoverflow.com/a/43154489">
-            2019 WhatsApp sharing standards (on StackOverflow)
-          </external-link>
-        </li>
-        <li>
-          <external-link href="https://stackoverflow.com/questions/19778620/provide-an-image-for-whatsapp-link-sharing">
-            Unfurl mechanism used by WhatsApp for sharing
-          </external-link>
-        </li>
-      </ul>
-    </panel-section>
   </div>
 </template>
 
 <script>
-import { computed } from 'vue';
+import { computed, onMounted, ref, watch } from 'vue';
 import createAbsoluteUrl from '@shared/lib/create-absolute-url';
-import { findMetaContent, findMetaProperty } from '@shared/lib/find-meta';
-import schema from '@shared/lib/schemas/whatsapp-schema';
+import {
+  findAdditionalTwitterData,
+  findFavicons,
+  findImageDimensions,
+  findMetaContent,
+  findMetaProperty
+} from '@shared/lib/find-meta';
+import getTheme from '@shared/lib/theme';
+import schema from './schema';
 
 import ExternalLink from '@shared/components/external-link';
 import PanelSection from '@shared/components/panel-section';
@@ -68,34 +61,40 @@ export default {
   },
 
   setup: props => {
+    const imageDimensions = ref({ height: undefined, width: undefined });
     const hasRequiredData = computed(() => (
       (og.value.title || props.headData.head.title) &&
-      (og.value.description || props.headData.headDescription)
+      (og.value.description || headDescription.value)
     ));
     const og = computed(() => ({
       title: propertyValue('og:title'),
-      description: propertyValue('og:description'),
       type: propertyValue('og:type'),
+      description: propertyValue('og:description'),
+      site_name: propertyValue('og:site_name'),
       image: propertyValue('og:image'),
       url: propertyValue('og:url'),
     }));
-    const title = computed(() => (
-      propertyValue('og:title') || props.headData.head.title || ''
+    const headDescription = computed(() => (findMetaContent(props.headData.head, 'description')));
+    const favicon = computed(() => (
+      findFavicons(props.headData.head).length ? findFavicons(props.headData.head)[0].url : ''
     ));
-    const description = computed(() => (og.value.description));
-    const hasDescription = computed(() => (
-      og.value.description !== null && og.value.description.length > 0
-    ));
-    const image = computed(() => (
-      og.value.image !== undefined ? absoluteUrl(og.value.image) : og.value.image
-    ));
+    const additional = computed(() => ({
+      favicon: favicon.value,
+      twitterData: findAdditionalTwitterData(props.headData.head),
+    }));
     const previewUrl = computed(() => {
       const params = new URLSearchParams();
-      params.set('title', og.value.title || title.value);
-      params.set('description', description.value);
-      params.set('image', image.value);
+      params.set('additionalData', JSON.stringify(additional.value.twitterData));
+      params.set('description', og.value.description || headDescription.value);
+      params.set('favicon', additional.value.favicon);
+      params.set('image', og.value.image);
+      params.set('imageIsBig', imageDimensions.value.height > 201 && imageDimensions.value.width > 201);
+      params.set('siteName', og.value.site_name);
+      params.set('theme', getTheme());
+      params.set('title', og.value.title || props.headData.head.title || 'Weblink');
       params.set('url', props.headData.head.url);
-      return `/previews/whatsapp/whatsapp.html?${ params }`;
+      params.set('validImage', imageDimensions.value.height > 0 && imageDimensions.value.width > 0);
+      return `/previews/slack/slack.html?${ params }`;
     });
     const metaData = computed(() => ([
       {
@@ -104,17 +103,21 @@ export default {
         required: true,
       },
       {
-        term: 'og:description',
-        value: og.value.description,
+        term: 'og:type',
+        value: og.value.type,
         required: true,
       },
       {
-        term: 'og:type',
-        value: og.value.type,
+        term: 'og:description',
+        value: og.value.description,
+      },
+      {
+        term: 'og:site_name',
+        value: og.value.site_name,
       },
       {
         term: 'og:image',
-        value: absoluteUrl(og.value.image),
+        value: og.value.image,
         image: {
           href: og.value.image,
           url: absoluteUrl(og.value.image),
@@ -124,23 +127,40 @@ export default {
       },
       {
         term: 'og:url',
-        value: og.value.url,
+        value: absoluteUrl(og.value.url),
         type: 'link',
         required: true,
       },
+      // Transform twitterData array and
+      // spread objects into metaData array.
+      ...additional.value.twitterData.map((item, index) => (
+        Object.entries(item).map(entry => ({
+          term: `twitter:${ entry[0] === 'label' ? entry[0] : 'data' }${ index + 1 }`,
+          value: entry[1],
+        }))
+      )).flat(),
     ]));
 
     const absoluteUrl = url => createAbsoluteUrl(props.headData.head, url);
+    const getImageDimensions = () => findImageDimensions(props.headData.head, 'og:image')
+      .then(dimensions => imageDimensions.value = dimensions);
     const propertyValue = propName =>
       findMetaProperty(props.headData.head, propName) || findMetaContent(props.headData.head, propName);
+
+    watch(() => og.value.image, value => {
+      if (value) {
+        getImageDimensions();
+      }
+    });
+
+    onMounted(() => getImageDimensions());
 
     return {
       hasRequiredData,
       og,
-      title,
-      description,
-      hasDescription,
-      image,
+      headDescription,
+      favicon,
+      additional,
       previewUrl,
       metaData,
       schema,
@@ -157,7 +177,7 @@ export default {
 </script>
 
 <style>
-  .whatsapp__preview {
+  .slack__preview {
     max-width: var(--preview-width);
   }
 </style>
