+++ projects/@shared/views/structured-data/structured-data.view.vue
